add sweet alert lib

+++ webapp/assets/js/publicacoes.js
@@ -20,7 +20,7 @@ function criarPublicacao(event) {
   }).done(function () {
     window.location = "/home";
   }).fail(function () {
-    alert("Erro ao criar publicação!");
+    Swal.fire("Ops...", "Erro ao criar publicação.", "error");
   })
 }
 
@@ -45,7 +45,7 @@ function curtirPublicacao(event) {
     click.removeClass('curtir-publicacao');
 
   }).fail(function () {
-    alert('Erro ao curtir publicação');
+    Swal.fire("Ops...", "Erro ao curtir publicação.", "error");
   }).always(function () {
     click.prop('disabled', false);
   });
@@ -73,7 +73,7 @@ function descurtirPublicacao(event) {
     click.addClass('curtir-publicacao');
 
   }).fail(function () {
-    alert('Erro ao curtir publicação');
+    Swal.fire("Ops...", "Erro ao descurtir publicação.", "error");
   }).always(function () {
     click.prop('disabled', false);
   });
@@ -92,9 +92,12 @@ function atualizarPublicacao() {
       conteudo: $('#conteudo').val()
     }
   }).done(function () {
-    alert("Publicação editada com sucesso")
+    Swal.fire("Sucesso!", "Publicação atualizada com sucesso!!", "success")
+      .then(function () {
+        window.location = "/home";
+      });
   }).fail(function () {
-    alert("Erro ao editar.")
+    Swal.fire("Ops...", "Erro ao editar", "error");
   }).always(function () {
     $('#atualizar-publicacao').prop('disabled', false);
   })
@@ -103,21 +106,30 @@ function atualizarPublicacao() {
 function deletarPublicacao(event) {
   event.preventDefault();
 
-  const click = $(event.target);
-  const publicacao = click.closest('div')
-  const publicacaoId = publicacao.data('publicacao-id');
+  Swal.fire({
+    title: "Atenção!",
+    text: "Tem certeza que deseja exlcuir essa publicação? Esta ação é irresersível!",
+    showCancelButton: true,
+    cancelButtonText: "Cancelar",
+    icon: "warning"
+  }).then(function (confirmation) {
+    if (!confirmation.value) return;
 
-  click.prop('disabled', true);
-  $.ajax({
-    url: `/publicacoes/${publicacaoId}`,
-    method: "DELETE"
-  }).done(function () {
-    publicacao.fadeOut("slow", function () {
-      $(this).remove();
+    const click = $(event.target);
+    const publicacao = click.closest('div')
+    const publicacaoId = publicacao.data('publicacao-id');
+
+    click.prop('disabled', true);
+    $.ajax({
+      url: `/publicacoes/${publicacaoId}`,
+      method: "DELETE"
+    }).done(function () {
+      publicacao.fadeOut("slow", function () {
+        $(this).remove();
+      });
+    }).fail(function () {
+      Swal.fire("Ops...", "Erro ao deletar a publicação.", "error");
     });
-  }).fail(function () {
-    alert("Erro ao deletar a publicação.");
-  });
-
+  })
 
 }
